Events edit/add - started

+++ imports/client/dangerous-room/event-list/event-list.component.ts
@@ -1,46 +1,48 @@
-import {Component, NgZone, OnInit} from '@angular/core';
+import { Component, NgZone, OnInit } from '@angular/core';
 import { Observable } from 'rxjs/Rx';
 import { TdDialogService } from '@covalent/core';
 
 import { DangerousRoomService } from '../shared/dangerous-room.service';
 
+import { BaseComponent } from '../../lib';
+
 import template from './event-list.component.html';
 import style from './event-list.component.scss';
 
 @Component({
-  styles: [style],
-  template
+    selector: 'dr-event-list',
+    styles: [style],
+    template
 })
-export class DREventComponent implements OnInit {
+export class DREventListComponent extends BaseComponent implements OnInit {
 
-  private event_list;
-  private showAddItemForm: boolean = false;
+    private event_list:EventItem[];
 
-  constructor (private _drService: DangerousRoomService,
-               private _dialogService:TdDialogService,
-               private _zone: NgZone
-  ) {
-  }
+    constructor (private _drService: DangerousRoomService,
+                 private _dialogService:TdDialogService,
+                 private _zone: NgZone
+    ) {
+        super();
+    }
 
-  ngOnInit(): void {
-    this._drService.allEvents$.subscribe((e) => {
-      this._zone.run(() => {
-          this.event_list = e;
-          console.log(e);
-      });
-    });
-  }
+    ngOnInit(): void {
+        this._drService.allEvents$.subscribe((e) => {
+            this._zone.run(() => {
+                this.event_list = e;
+                console.log(e);
+            });
+        });
+    }
 
-  onDeleteItem(itemID: string): void {
-    this._drService.deleteEvent(itemID);
-  }
+    onDeleteItem(itemID: string): void {
+        this._dialogService.openConfirm({
+            "title": "Delete an event?",
+            "message": "Please confirm!"
 
-  onHideAddEventForm(value: boolean): void {
-    this.showAddItemForm = value;
-  }
-
-  onShowAddEventForm(): void {
-    // this._dialogService.openAlert({title:'test',message:'Form to show'});
-    this.showAddItemForm = true;
-  }
+        }).afterClosed().subscribe((accept: boolean) => {
+            if (accept) {
+                this._drService.deleteEvent(itemID);
+            }
+        });
+    }
 }
